tests for book details moved to book details.
